+++ Magic/Framework/PreparedSpell.cs
@@ -4,9 +4,19 @@ namespace Magic.Framework
 {
     internal class PreparedSpell
     {
+        /*********
+        ** Accessors
+        *********/
+        /// <summary>The learned spell ID.</summary>
         public string SpellId { get; set; }
+
+        /// <summary>The spell level.</summary>
         public int Level { get; set; }
 
+
+        /*********
+        ** Public methods
+        *********/
         public PreparedSpell() { }
 
         public PreparedSpell(string spellId, int level)